src/trypto/exchange_manager.py
import asyncio
import aiohttp
import websockets
import logging
import time
from collections import defaultdict
from retry_policy import RetryPolicy
from rate_limiter import RateLimiter
from connection_health import ConnectionHealth
from ws_feed_handler import WSFeedHandler
from typing import List, Dict, Any

class ExchangeManager:
    def __init__(self):
        self.connections = {}
        self.rate_limiters = {}
        self.websocket_feeds = {}
        self.exchange_configs = self._load_exchange_configs()
        self.connection_health = {}
        self.retry_policy = RetryPolicy(
            max_retries=3,  # Changed from max_attempts
            base_delay=1.5  # Changed from backoff_factor
        )
        self.fallback_exchanges = ['binance', 'ftx', 'coinbase']
        self.current_exchange_index = 0
        self.positions = {}
        self.orders = {}

    async def initialize(self):
        """Initialize exchange connections and configurations"""
        try:
            # Initialize exchange connections
            # This is a placeholder - add actual exchange initialization logic
            return True
        except Exception as e:
            logging.error(f"Failed to initialize ExchangeManager: {str(e)}")
            raise

    async def close_all_positions(self):
        """Close all open positions."""
        try:
            # Mock implementation
            self.positions = {}
            return True
        except Exception as e:
            self.logger.error(f"Failed to close positions: {str(e)}")
            return False

    async def cancel_all_orders(self):
        """Cancel all pending orders."""
        try:
            # Mock implementation
            self.orders = {}
            return True
        except Exception as e:
            self.logger.error(f"Failed to cancel orders: {str(e)}")
            return False

    async def shutdown(self):
        """Cleanup and close exchange connections"""
        try:
            # Close positions and cancel orders
            await asyncio.gather(
                self.close_all_positions(),
                self.cancel_all_orders(),
                return_exceptions=True
            )

            # Close websocket connections
            for feed in self.websocket_feeds.values():
                if hasattr(feed, 'close'):
                    await feed.close()
            
            # Close HTTP sessions
            for session in self.connections.values():
                if hasattr(session, 'close'):
                    await session.close()
            
            # Clear all data structures
            self.connections.clear()
            self.websocket_feeds.clear()
            self.rate_limiters.clear()
            self.connection_health.clear()
            
            return True
        except Exception as e:
            logging.error(f"Failed to shutdown ExchangeManager: {str(e)}")
            return False

    def _load_exchange_configs(self):
        # Load exchange configurations from a file or environment variables
        return {
            'binance': {
                'rest_endpoint': 'https://api.binance.com',
                'ws_endpoint': 'wss://stream.binance.com:9443/ws',
                'rate_limits': {'orders': 1200, 'window': 60}
            },
            'coinbase': {
                'rest_endpoint': 'https://api.exchange.coinbase.com',
                'ws_endpoint': 'wss://ws-feed.exchange.coinbase.com',
                'rate_limits': {'orders': 500, 'window': 60}
            },
            # Add other exchange configurations as needed
        }

    async def initialize_exchange(self, exchange_id, credentials):
        try:
            config = self.exchange_configs.get(exchange_id)
            if not config:
                raise ValueError(f"Unsupported exchange: {exchange_id}")

            connection = await self.retry_policy.execute(
                self._create_connection,
                config['rest_endpoint'],
                credentials
            )

            self.connection_health[exchange_id] = ConnectionHealth(
                heartbeat_interval=30,
                timeout_threshold=5
            )

            self.connections[exchange_id] = connection
            self.rate_limiters[exchange_id] = RateLimiter(
                config['rate_limits'],
                burst_allowance=1.2
            )

            self.websocket_feeds[exchange_id] = await self._setup_websocket(
                exchange_id,
                config['ws_endpoint'],
                keepalive_interval=30
            )

            asyncio.create_task(self._monitor_connection_health(exchange_id))

        except Exception as e:
            logging.error(f"Failed to initialize exchange {exchange_id}: {str(e)}")
            raise

    async def _create_connection(self, rest_endpoint, credentials):
        timeout = aiohttp.ClientTimeout(total=10)
        connector = aiohttp.TCPConnector(limit_per_host=10)
        session = aiohttp.ClientSession(timeout=timeout, connector=connector)
        # Implement authentication with credentials
        if credentials:
            session.headers.update({'Authorization': f"Bearer {credentials['api_key']}"})
        return session

    async def _setup_websocket(self, exchange_id, endpoint, keepalive_interval):
        async def reconnect():
            while True:
                try:
                    ws = await websockets.connect(endpoint, ping_interval=keepalive_interval)
                    feed_handler = WSFeedHandler(ws, exchange_id)
                    asyncio.create_task(feed_handler.listen())
                    return feed_handler
                except Exception as e:
                    logging.error(f"WebSocket connection error for {exchange_id}: {str(e)}")
                    await asyncio.sleep(5)

        return await reconnect()

    async def _monitor_connection_health(self, exchange_id):
        while True:
            is_healthy = await self.connection_health[exchange_id].check_health()
            if not is_healthy:
                logging.warning(f"Connection health degraded for {exchange_id}. Attempting to reconnect.")
                await self.initialize_exchange(exchange_id, credentials={})
            await asyncio.sleep(self.connection_health[exchange_id].heartbeat_interval)

    async def api_call(self, method: str, *args, **kwargs) -> Any:
        while True:
            exchange = self.fallback_exchanges[self.current_exchange_index]
            
            if await self.rate_limiter.acquire(exchange):
                try:
                    return await self.retry_policy.execute(
                        self._make_api_call,
                        exchange,
                        method,
                        *args,
                        **kwargs
                    )
                except Exception as e:
                    # Try next exchange in fallback list
                    self.current_exchange_index = (self.current_exchange_index + 1) % len(self.fallback_exchanges)
                    if self.current_exchange_index == 0:
                        raise  # We've tried all exchanges
            
            await asyncio.sleep(0.1)  # Prevent tight loop

    # ...rest of implementation...